fixed all the tests
