+++ app/api/validation/validatorFactory.ts
@@ -3,14 +3,6 @@ import * as ajvasync from "ajv-async";
 
 import { DateValidator } from "../validation/dateValidator";
 
-// Test reqs
-import "mocha";
-import "mocha-inline";
-import * as chai from "chai";
-import * as chaiAsPromised from "chai-as-promised";
-import * as TypeMoq from "typemoq";
-import { Fixtures } from "../test/fixtures";
-import { ProductValidator } from "../validation/productValidator";
 
 export function getValidator(schemaVersion:String = "", options: any = { allErrors: true, formats: "full" }) {
   let validator = ajv(options);
@@ -41,26 +33,3 @@ export function getValidator(schemaVersion:String = "", options: any = { allErro
 
   return ajvasync(validator);
 }
-
-// Tests
-// Test setup
-// tslint:disable-next-line:no-var-requires
-chai.use(chaiAsPromised);
-
-describe("Legacy json schema spec support", () => {
-  it("should support draft-04 spec schemas", () => {
-    let mockRepo = Fixtures.GetMockCollectionStore(4);
-    let validator = new ProductValidator(mockRepo.object);
-    return chai.expect(validator.validate(Fixtures.GetTestProduct()))
-      .to.be.fulfilled
-      .and.eventually.be.an("array").that.is.empty;
-  });
-
-  it("should support draft-06 spec schemas", () => {
-    let mockRepo = Fixtures.GetMockCollectionStore(6);
-    let validator = new ProductValidator(mockRepo.object);
-    return chai.expect(validator.validate(Fixtures.GetTestProduct()))
-      .to.be.fulfilled
-      .and.eventually.be.an("array").that.is.empty;
-  });
-});
